+++ esperanto-tools-libs/test_package/test_package.cpp
@@ -1,6 +1,7 @@
 #include "runtime/IProfileEvent.h"
 #include "runtime/IProfiler.h"
 #include "runtime/IRuntime.h"
+#include "runtime/Types.h"
 
 #include <iostream>